Add files via upload
I recently wrote some code to practice multidimensional arrays in Java. As I was reflecting on what to implement, I realized that I am currently pursuing a minor in mathematics alongside my major in computer science. This semester, I’ve been studying linear algebra, which included a review of matrix concepts.

I thought, “Why not try to implement that in my code?” This not only would enhance my understanding of loops but also help me improve my skills with multidimensional arrays. And here we are!